Adapt Conduit to use new phly/http mutator verbiage
Instead of `setters`, `with*()` methods. Additionally, requests now only
compose a Uri instance, which means we do not need to use `parse_url()`
and/or casting to a URI to get at the URI segments.

This also exposed a flaw in the logic of Next for truncating paths,
which is now fixed.