[phly/conduit#25] Fix resetting requests re: trailing slashes
Added another test case to test for the inverse condition -- that a
trailing slash **is** present if it was present in the original route.

The fix requires ensuring that the original route is retained, but that
matching is done without trailing slashes. At the same time, the
original route must be retained so that the path is reset correctly --
without double slashes.