Created interfaces for describing middleware
MiddlewareInterface, to describe normal middleware, and
ErrorMiddlewareInterface, to describe middleware for handling errors.
Internally, Dispatch will use these to determine the type of middleware
being executed (and fall back to arity if the middleware does not
implement one of them).

By doing this, I also was able to remove the logic for using a "handle"
method, as it no longer needs to be supported.

I've now written all the various signatures Next() supports; I'm leaning
towards a required signature at this time, as it's getting too
complicated.