Refactored to a consistent Next() signature
`Next()` now ALWAYS expects a request and response. This obviates the
need to store them internally, and simplifies the logic tremendously
both for implementation as well as consumption -- though ALL calls to
request now need to pass the request and response, minimally.

Additionally, since this operates as a queue, it made sense to use
SplQueue. This affected a few tests, but does not affect the public API
in the least. Next was altered to use the verbiage queue instead of
stack (missed that on a previous commit).