Added more nesting path test cases
Added variants to each existing test with the suffix "plus", which adds
an additional word to the end of the path to test for matching. In some
situations, this MUST NOT match (e.g., `/foo/barbaz` should not match
against nested `/foo` and `/bar` middlewares), and in others, it should
(e.g., `/foo/bar/baz` should match those same middlewares).