FinalHandler should return new responses
If no error is present, and the response passed on invocation is
different than the response at initialization, the `FinalHandler` should
return that response intact.

This allows middleware to create a response, but pass it on to `$next()`
for post-processing further up the chain. In particular, this is useful
for things like routing middleware, so as to allow for application-wide
concerns such as logging, finalizing sessions, etc., without requiring
onion-style middleware for these behaviors.