Allow PSR-7 messages to Dispatch
`Next` allows any PSR-7 implementations. However, prior to this patch,
`Dispatch` required the Stratigilty-specific decorators, which could
result in issues if `$next()` was invoked with non-Stratigility
instances:

```php
return $next($request, new JsonResponse());
```

This patch adds a test to ensure that `Dispatch` will pass them through
properly, and modifies the typehints to use the PSR-7 interfaces. The
change is completely BC, as the Stratigility implementations are a
superset of PSR-7, and no specfic functionality from them was used
previously.