Fixed failing test
Patch #39 added the method `testInvokingWithExceptionInNonProductionModeIncludesPrevTraceInResponseBody`.
The test fails after adding the code from #46, as that code modifies
`FinalHandler` to never provide exception details unless in
non-production mode (which is the default). This patch updates the test
to set the final handler used in the test setup to set the environment
to development so that the expectations pass.