Fix "500 OK" error responses
If the current response already has a status code that corresponds to
the exception that was encountered, it is assumed the developer dealt
with the reason phrase as well, and the reason phrase will be preserved.
If not, the standard message for the resulting HTTP status code will be
used.