Do not mark response as complete in FinalHandler
As reported in #53, because utilities such as the `SapiEmitter` need to update
the response prior to sending it, marking the response as complete has an
adverse impact since the release of 1.2.0, as this will now cause an exception.

This patch fixes that situation by using the `write()` method in favor of
`end()` within the `FinalHandler::completeResponse()` method.