Added Dispatch::process to allow processing http-interop middleware
`Next` will be called via its `process()` method when invoked by
http-interop middleware; as such, it can invoke a separate method on
Dispatch.

This patch introduces `Dispatch::process()`, which then determines how
to delegate the route provided. For non-http-interop middleware, it will
raise an exception if the `Dispatch` instance does not have a response
prototype injected, or if the request is a non-server-side request. When
dispatching http-interop middleware, if an exception is caught, it
raises exceptions for the same conditions, but otherwise delegates to
the next error handler.