Ensure Next processes the next delegate correctly when an error is present
Discovered when updating tests for Expressive to work with Stratigility
1.3. Essentially, when nesting pipelines, if an inner middleware
pipeline is exhausted and a non-null `$err` is present, if the next
delegate was a `Next` instance, it was improperly being called as an
http-interop `DelegateInterface`, losing the error argument. As such, if
any parent delegate contained error middleware, it would never be
invoked.