doc-block typo
Hello @weierophinney,
wouldn't it also make more sense to rename the `$delegate` property and parameter to `$next` given that the purpose of this decorator is to wrap a standard - stratigility < 1.3 - $next callable? I visually (and mentally) tend to associate the `$delegate` job with `$delegat->process($req)` calls and old `$next` callables job with `$next($req, $res)` invokations. In this case the decorator "decorates the callable" by "adding" its own process() method with a proxy invokation of $next.

kind regards