Ensure that MiddlewarePipes are called as callables if an error is present
As reported in zendframework/zend-expressive#416, error middleware
nested inside a `MiddlewarePipe` was not being dispatched. This was due
to the fact that `Dispatch` was identifying the pipeline as http-interop
middleware, and thus dropping the `$err` argument (as interop middleware
cannot accept that argument).

Theis patch updates `Dispatch` to check if the middleware is a
`MiddlewarePipe` and a non-null `$err` is present; if so, it now
dispatches it as callable middleware instead of as interop middleware.