Pipe should not decorate MiddlewarePipe instances
In prepping Expressive to use Stratigility 2, I discovered that
`MiddlewarePipe` instances piped into another `MiddlewarePipe` instance
were being decorated as callable middleware; this should no longer
happen.

This patch adds a test to ensure it does not happen, and updates
`pipe()` to not decorate callable `ServerMiddlewareInterface` instances.

In doing so, the `isInteropMiddleware()` method is no longer relevant,
and I removed it; since it was marked `private`, this is a backwards
compatible change.