Adds PSR-15 decorator classes for callable middleware
This patch provides two classes, `CallableMiddlewareDecorator` and
`DoublePassMiddlewareDecorator`. The former replaces the previous
`CallableInteropMiddlewareWrapper`, and the latter is the new version of
`CallableMiddlewareWrapper` from the v2 series.

`CallableMiddlewareDecorator` accepts a callable that is expected to
conform to the PSR-15 signature basics (accepts a request and handler,
and returns a response), but does no verifications on the signature. If
no response is returned by the callable, it raises an exception.

`DoublePassMiddlewareDecorator` accepts a callable that follows the
double-pass signature (`function ($request, $response, $next)`),
but does no verifications on the signature. If no response is returned
by the callable, it raises an exception.