Implements a PathMiddlewareDecorator
This replaces the path segregation features of the original stratigility
MiddlewarePipe with a middleware decorator.

The decorator receives the path prefix to segregate by, and the
middleware to dispatch when matched.

During `process()`, it checks to see if the prefix is matched by the
current URI path at an appropriate boundary. If not, it acts as a
pass-through to the handler.

If it *does* match, it does the following:

- Creates a new request that contains a URI that truncates the prefix
  from the path.
- Prepares a request handler that decorates the original in order to
  process it with the original incoming request.
- Dispatches the composed middleware using the new request and the
  decorated request handler.

I have pulled relevant tests from the original test suite related to
path segregation and adapted them to test against the new
PathMiddlewareDecorator; I have also added an integration test
demonstrating that nesting works as expected within a pipeline.