Deprecates path segregation via MiddlewarePipe::pipe()
This patch deprecates using the two-argument form of
`MiddlewarePipe::pipe()` in favor of using a `PathMiddlewareDecorator`
instance instead. It accomplishes this via the following:

- When a non-empty, non-root path argument is provided along with
  middleware, `MiddlewarePipe::pipe()` will trigger an
  `E_USER_DEPRECATED` error to notify the user that path segregation
  should be accomplished using the `PathMiddlewareDecorator`.
- When `MiddlewarePipe::pipe()` is called with two arguments, the
  path argument is not a root path, and the middleware is not a
  `PathMiddlewareDecorator` instance, it will be decorated as one.
- The constructor of `Route` will trigger an `E_USER_DEPRECATED` error
  if a non root path argument is provided with a
  non-`PathMiddlewareDecorator` middleware instance. It will then
  decorate the middleware in a `PathMiddlewareDecorator` instance.
- `Next` no longer handles path segregation internally.

The changes to `Next` are _almost_ a BC break. However, it includes some
code to cast `Route` instances that represent path-segregated middleware
to `PathMiddlewareDecorator` instances in the case that they are not
already (which can only happen with custom extensions). As such,
behavior is retained with this patch.

In updating tests, I discovered some edge cases to using the
`PathMiddlewareDecorator` approach; in particular, a path-segregated
middleware should be able to propagate changes to the request when
calling the handler passed to it. As such, the
`PathRequestHandlerDecorator` now updates the request URI with the path
from the original request URI only.