Trigger deprecation notices when piping callable middleware
This patch updates the logic in `MiddlewarePipe` as follows:

- Detection of callable middleware, either interop or double-pass, now
  triggers an `E_USER_DEPRECATED` error indicating that this feature is
  deprecated, and that users should decorate their middleware before
  passing it to `pipe()`, recommending the backported decorator classes.
- By default, decoration of interop middleware is now done with a
  `CallableMiddlewareDecorator` instance.
- By default, decoration of double-pass middleware is now done with a
  `DoublePassMiddlewareDecorator` instance, unless a
  `CallableMiddlewareWrapperFactory` is composed.

Tests were refactored to decorate callable middleware when not testing
the decoration features. When testing the decoration features, tests
were updated to detect emission of the deprecation notices.

The `RouteTest` was updated to follow the same patterns as used in the
`MiddlewarePipeTest` for detection of the deprecation notice in order to
remove the necessity to close over a reference.