Ensures that path segregated middleware may provide an altered request
Discovered while backporting to the 2.X series. Essentially, a path
segregated middleware should still be able to pass a new request on to
the next layer. However, in doing so, the _path_ should be reset.

This patch updates the `PathMiddlewareDecoratorIntegrationTest` to
reflect this behavior, and updates the anonymous request handler created
by the `PathMiddlewareDecorator` to invoke the composed handler with the
runtime request after resetting its URI instance with the original path.