Adds redirect pages for legacy URLs
A new feature in the zf-mkdoc-theme allows you to mark a page or section
to hide from navigation by prefixing the title with an underscore; the
page content is then rendered only in the `<head>` of the document,
allowing us to define meta refresh and JS redirect routines.

This patch does so for all pages in the currently released
documentation, allowing the legacy URLs to now redirect to the v3
variants.