Creates the `RequestHandlerMiddleware` class
This class will decorate a `RequestHandlerInterface` instance such that
it can act as either a request handler or middleware. When either
`handle()` or `process()` are called, the class will instead proxy to
the composed handler's `handle()` method, returning the result.

This can be useful for decorating request handlers to use within
pipelines.