Backport fixes from #165 to v2.2 release
Updates the `PathRequestHandlerDecorator` to accept a string prefix
instead of a `ServerRequestInterface` to its constructor, and then to
use that to build the request path to pass to the handler. The
`PathMiddlewareDecorator` now passes the prefix instead of the original
request to the `PathRequestHandlerDecorator` constructor.

New test passes, but one existing test fails at this point.