Fixed invalid test setup in NextTest
The test testMiddlewareCallingNextWithRequestPassesRequestToNextMiddleware
was using path segregated routes, but they were not pertinent to the
behavior being tested, and, in fact, broke once we fixed the behavior
reported in #165. The equivalent version 3 tests do not use path
segregated routes, so the path segregation was removed in order to
test the expected behavior correctly.