Fixes path truncation in PathMiddlewareDecorator to be case insensitive
In #168, we received a report indicating that path matching within
`PathMiddlewareDecorator` fails if the path is the same, but using a
different case.

Interestingly, we _match_ it correctly; the problem is when _preparing
the truncated path_ to pass on to the decorated middleware, as it cannot
identify the matched segment correctly.

This patch modifies the logic in `getTruncatedPath()` to do a
case-insensitive comparison of the segment and the path in the initial
condition.